added cleanup script
