+++ tests/validation_v2/storage/scripts/common_fixtures.py
